refactr(client): remove unused code; move files/folders; lint/format fix
