--- web/resources/js/Components/ApplicationLogo.vue
@@ -1,3 +1,0 @@
-<template>
-  <img :src="$page.props.logo" alt="mecode" />
-</template>
