--- web/resources/js/Components/ApplicationMark.vue
@@ -1,3 +1,0 @@
-<template>
-  <img :src="$page.props.logo" alt="mecode" />
-</template>
