+++ web/resources/js/Pages/CodingPage.bak.vue
@@ -734,7 +734,7 @@ import {
 //import { Switch, SwitchGroup, SwitchLabel } from '@headlessui/vue';
 // import { closeFullscreen, openFullscreen } from '../utils/dom/fullScreen.js';
 import { XCircleIcon as XCircleSolidIcon } from '@heroicons/vue/24/solid';
-import { vClickOutside } from '../Components/coding/clickOutsideDirective.js';
+import { vClickOutside } from '../utils/vue/clickOutsideDirective.js';
 // import DropdownMenu from '../Components/coding/DropdownMenu.vue';
 // import {
 //     ArrowsPointingInIcon,
