+++ web/resources/js/Pages/ProjectsList.vue
@@ -11,7 +11,7 @@ import AuthenticatedLayout from '../Layouts/AuthenticatedLayout.vue';
 import ProjectsListMenu from './Projects/ProjectsListMenu.vue';
 import BaseContainer from '../Layouts/BaseContainer.vue';
 import CreateDialog from '../dialogs/CreateDialog.vue';
-import { useProjects } from './Projects/useProjects.js';
+import { useProjects } from '../domain/project/useProjects.js';
 import HelpResources from '../Components/HelpResources.vue';
 import Headline2 from '../Components/layout/Headline2.vue';
