+++ web/resources/js/Pages/coding/contextMenu/CodingContextMenu.vue
@@ -8,11 +8,11 @@ import Button from '../../../Components/interactive/Button.vue';
 import { cn } from '../../../utils/css/cn';
 import { ref, computed, watch } from 'vue';
 import InputField from '../../../form/InputField.vue';
-import { vClickOutside } from '../../../Components/coding/clickOutsideDirective';
+import { vClickOutside } from '../../../utils/vue/clickOutsideDirective';
 import CodingContextMenuItem from './CodingContextMenuItem.vue';
 import { useSelections } from '../selections/useSelections';
 import { useContextMenu } from './useContextMenu';
-import { useCodes } from '../useCodes';
+import { useCodes } from '../../../domain/codes/useCodes';
 import { useRange } from '../useRange';
 import { whitespace } from '../../../utils/regex';
 import { useUsers } from '../../../domain/teams/useUsers';
